requirements.txt
Django
djangorestframework
ebaysdk

numpy
pandas
nltk

pytest
pytest-django